Initial working state
