+++ core/__init__.py
@@ -1,6 +1,7 @@
 from core.config import Config
 from subprocess import run, PIPE
 from collections import namedtuple
+from core.cli import determine_main_folder
 
 __all__ = ["Config", "__version__"]
 version_info = namedtuple("VersionInfo", "major minor patch")
@@ -10,8 +11,11 @@ BASE_VERSION = version_info(3, 0, 0)
 
 def get_latest_version():
     try:
+        cmd = "cd {} && git describe --abbrev=0 --tags".format(
+            determine_main_folder()
+        )
         p = run(
-            "git describe --abbrev=0 --tags".split(),
+            cmd.split(),
             stdout=PIPE
         )
     except FileNotFoundError:
